backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.12.5
pydantic-settings==2.6.1
python-multipart==0.0.21
langgraph==1.0.10
google-cloud-aiplatform==1.134.0
google-cloud-firestore==2.27.0
